+++ python/seqpro/rag/__init__.py
@@ -1,5 +1,5 @@
 from ._array import DTYPE_co, Ragged, RDTYPE_co, is_rag_dtype
-from ._ops import reverse_complement, to_packed
+from ._ops import reverse_complement, to_packed, to_padded
 from ._utils import OFFSET_TYPE, lengths_to_offsets
 
 __all__ = [
@@ -11,4 +11,5 @@ __all__ = [
     "lengths_to_offsets",
     "reverse_complement",
     "to_packed",
+    "to_padded",
 ]